+++ valinta-tulos-service/src/main/scala/fi/vm/sade/valintatulosservice/config/EmailerConfig.scala
@@ -1,4 +1,4 @@
-package fi.vm.sade.valintatulosemailer.config
+package fi.vm.sade.valintatulosservice.config
 
 import java.util.concurrent.TimeUnit.SECONDS
