+++ valinta-tulos-service/src/test/scala/fi/vm/sade/valintatulosservice/vastaanottomeili/EmailerServletSpec.scala
@@ -1,17 +1,8 @@
-package fi.vm.sade.valintatulosemailer
+package fi.vm.sade.valintatulosservice.vastaanottomeili
 
 import fi.vm.sade.valintatulosservice.ServletSpecification
-import fi.vm.sade.valintatulosservice.json.JsonFormats
-import fi.vm.sade.valintatulosservice.valintarekisteri.ValintarekisteriDbTools
-import fi.vm.sade.valintatulosservice.vastaanottomeili.{LahetysKuittaus, PollResult}
-import org.json4s._
-import org.json4s.jackson.Serialization
-import org.json4s.native.JsonMethods._
 import org.junit.runner.RunWith
-import org.specs2.matcher.MatchResult
 import org.specs2.runner.JUnitRunner
-import org.specs2.specification.BeforeEach
-import slick.jdbc.PostgresProfile.api._
 
 
 @RunWith(classOf[JUnitRunner])
